lightborne
Module camera

Structs

Animation

BackgroundCamera

Marker Component used to query for the background camera. Note that for an entity to be rendered on this Camera, it must be given the RenderLayers::layer(1) component.

CameraPlugin

The Plugin responsible for handling anything Camera related.

MainCamera

Marker Component used to query for the main camera in the world.

SetStatePlayingCallback

Enums

MoveCameraEvent

Constants

CAMERA_ANIMATION_SECS 🔒

CAMERA_HEIGHT 🔒

CAMERA_WIDTH 🔒

Functions

handle_move_camera

move_camera

System that moves camera to player’s position and constrains it to the CurrentLevel’s world_box.

set_state_playing

setup_camera 🔒

Startup System that spawns the Camera2d in the world.