lightborne::level
Module activatable

Structs

Activatable

Component that shuold be inserted into any entity that is activatable. The ActivatablePlugin will manage each activatable’s state by listening to GroupTriggeredEvents, and inserting Activated components on the entities, if they should be activated. Each entity can then listen for changes with Added or RemovedComponents, and apply their logic in their corresponding System.

ActivatableCache

Resource used to store the entity handles of each Activatable by their id, for faster performance.

ActivatablePlugin

Plugin that manages all activatables in the game, e.g. LightSensor. See Activatable for more information.

Activated

Component that indicates whether or not an entity should be activated or not. See Activatable for more information.

GroupTriggeredEvent

Event that should be sent from an Interactable to notify this plugin to add or remove Activated markers.

Functions

init_activatable

Initialization function to be used to intialize Activatable LDTK entities. See crate::level::crystal::CrystalBundle for an example of how this can be done.

setup_activatables

When Activatables are spawned, we need to insert the Activated component on entities that should be intially active. This is where that is done.

update_activatables

The Update System responsible for responding to GroupTriggeredEvents, and inserting Activated markers.