lightborne::level::activatable
Function init_activatable

pub fn init_activatable(entity_instance: &EntityInstance) -> Activatable

Initialization function to be used to intialize Activatable LDTK entities. See crate::level::crystal::CrystalBundle for an example of how this can be done.