lightborne::level
Module crystal

Structs

Crystal

Marker Component used to query for crystals, currently does not contain any information.

CrystalBundle

Bundle registered with LdktEntityAppExt::register_ldtk_entity to spawn crystals directly from Ldtk.

CrystalCache

CrystalColor

Enum that represents the crystals that a [LightSensor] should toggle. Differs from the LightColor in that the white color requires an ID field.

CrystalId

Identifier Component used to label the ID of white crystals

CrystalIdBundle

Bundle registered with LDTK to spawn in white crystal identifiers

CrystalPlugin

Plugin for managing all things related to Crystals. This plugin responds to the addition and removal of [Activated] Components and updates the sprite and collider of each crystal entity, in addition to handling initialization and cleanup on a [LevelSwitchEvent].

CrystalToggleEvent

Event that will toggle all crystals of a certain color.

Constants

CRYSTAL_INDEX_OFFSET 🔒

The horizontal offset between active crystals and inactive crystals in the crystal tilemap

Functions

activate_crystal 🔒

Switches a crystal from inactive to active. Calling this on an already active crystal will result in weird behavior.

add_crystal_colliders 🔒

crystal_color 🔒

Function to determine the base color of the crystal.

deactivate_crystal 🔒

Switches a crystal from active to inactive. Calling this on an already inactive crystal will result in weird behavior.

init_crystal_cache_and_ids 🔒

System that will initialize all the crystals, storing their entities in the appropriate level -> crystal color location in the crystal cache.

is_crystal_active 🔒

Function to determine whether or not a cell value represents an Active Crystal. Does not use the modulo operator as future crystal cell values need not necessarily follow the same pattern in the future.

on_crystal_changed

System that listens to when Crystals are activated or deactivated, updating the Sprite and adding/removing [FixedEntityBundle] of the Entity.

reset_crystals

System that listens to [LevelSwitchEvent]s to ensure that Crystal states are reset when switching between rooms.

update_crystal_cache 🔒