lightborne::level::crystal
Function on_crystal_added

pub fn on_crystal_added(
    commands: Commands<'_, '_>,
    q_new: Query<'_, '_, (Entity, &mut Sprite, &Activatable), Added<Crystal>>,
)

System to ensure that the sprite color and collider match the initially activated state of the Crystal.