lightborne::level::crystal
Function on_crystal_changed

pub fn on_crystal_changed(
    commands: Commands<'_, '_>,
    q_activated: Query<'_, '_, Entity, (With<Crystal>, Added<Activated>)>,
    q_deactivated: RemovedComponents<'_, '_, Activated>,
    q_crystals: Query<'_, '_, (Entity, &mut Sprite, &EntityInstance), With<Crystal>>,
)

System that listens to when Crystals are activated or deactivated, updating the Sprite and adding/removing FixedEntityBundle of the Entity.

We should instead consider inserting a Sensor component and updating the entity’s CollisionGroups instead, if we are worried about the entity changing its Archetype and causing performance issues (not likely, Crystals do not change state often).