lightborne::level::walls
Function spawn_wall_collision

pub fn spawn_wall_collision(
    commands: Commands<'_, '_>,
    wall_query: Query<'_, '_, (&GridCoords, &Parent), Added<Wall>>,
    parent_query: Query<'_, '_, &Parent, Without<Wall>>,
    level_query: Query<'_, '_, (Entity, &LevelIid)>,
    ldtk_projects: Query<'_, '_, &LdtkProjectHandle>,
    ldtk_project_assets: Res<'_, Assets<LdtkProject>>,
)

NOTE: copied from the bevy_ecs_ldtk examples

Spawns heron collisions for the walls of a level

You could just insert a ColliderBundle into the WallBundle, but this spawns a different collider for EVERY wall tile. This approach leads to bad performance.

Instead, by flagging the wall tiles and spawning the collisions later, we can minimize the amount of colliding entities.

The algorithm used here is a nice compromise between simplicity, speed, and a small number of rectangle colliders. In basic terms, it will:

consider where the walls are

combine wall tiles into flat “plates” in each individual row

combine the plates into rectangles across multiple rows wherever possible

spawn colliders for each rectangle