lightborne::light
Module segments

Structs

LightBounceSfx

Local variable for simulate_light_sources used to store the handle to the audio SFX

LightSegmentBundle

Bundle used in the initialization of the LightSegmentCache to spawn segment entities.

LightSegmentCache

Resource used to store Entity handles to the light segments so they aren’t added and despawned every frame. See simulate_light_sources for details.

LightSegmentMarker

Marker Component used to query for light segments.

Functions

cleanup_light_sources

System that is responsible for hiding all of the LightSegments and despawning LightRaySources when the level changes.

simulate_light_sources

System that runs on Update, calculating the Transform of light segments from the corresponding LightRaySource. Note that this calculation happens every frame, so instead of rapidly spawning/despawning the entities, we spawn them and cache them in the LightSegmentCache, then modify their Visibility and Transforms.

tick_light_sources

System that runs on FixedUpdate, advancing the distance the light beam can travel.