lightborne::light::segments
Function simulate_light_sources

pub fn simulate_light_sources(
    commands: Commands<'_, '_>,
    q_light_sources: Query<'_, '_, &mut LightRaySource>,
    q_rapier: Query<'_, '_, &mut RapierContext>,
    ev_hit_by_light: EventWriter<'_, HitByLightEvent>,
    q_light_sensor: Query<'_, '_, &LightSensor>,
    q_segments: Query<'_, '_, (&mut Transform, &mut Visibility), With<LightSegmentMarker>>,
    segment_cache: Res<'_, LightSegmentCache>,
    light_bounce_sfx: Local<'_, LightBounceSfx>,
)

System that runs on Update, calculating the Transform of light segments from the corresponding LightRaySource. Note that this calculation happens every frame, so instead of rapidly spawning/despawning the entities, we spawn them and cache them in the LightSegmentCache, then modify their Visibility and Transforms.

If needed, optimization work can be done by recalculating only segments that are currently changing (segments already “stabilized” usually won’t move).

Similar logic is duplicated in preview_light_path, these two systems should be merged.