lightborne::light
Module sensor

Structs

HitByLight

Component used to mark components that have been hit by light. The current design of the system is very bad, an event like HitByLightEvent(Entity) should be used instead to signal the LightSensors to activate.

LightSensor

Component added to entities receptive to light. The activation_timer should be initialized in the From<&EntityInstance> implemenation for the LightSensorBundle, if not default.

LightSensorBundle

Bundle that includes all the Components needed for a LightSensor to function properly.

Functions

reset_light_sensors

System that resets the LightSensors when a LevelSwitchEvent is received.

update_light_sensors

System that queries LightSensors for HitByLight markers added when light hits a light sensor, and immediately removes them. As mentioned in HitByLight, this design pattern isn’t the best, and the Event-based one should be implemented instead.