lightborne::light::sensor
Function update_light_sensors

pub fn update_light_sensors(
    commands: Commands<'_, '_>,
    q_sensors: Query<'_, '_, (Entity, &mut LightSensor)>,
    ev_hit_by_light: EventReader<'_, '_, HitByLightEvent>,
    ev_crystal_toggle: EventWriter<'_, CrystalToggleEvent>,
    asset_server: Res<'_, AssetServer>,
    time: Res<'_, Time>,
)

System that runs on Update, querying each LightSensor and updating them based on each HitByLightEvent generated in the System: simulate_light_sources. This design is still imperfect, as while it differs semantically from the previous implementation, each Event is generated every frame. Preferably, refactor to include a “yap”-free implementation across multiple systems to better utilize Event.