lightborne::lighting
Module occluder

Structs

ColliderBasedOccluder

Component that automatically attaches Occluder components as children of an entity based on the shape of its Collider. Also automatically removes and reattaches Occluders when the Collider is removed/reattached (e.g. crystals). Currently only works on cuboid colliders.

Occluder

Component that represents a line that prevents light from passing through. The two points that make up the line are calculated by adding point_1_offset and point_2_offset to the entity’s Transform.

OccluderPlugin

OccluderRenderer 🔒

Component used to internally to render each occluder. Every time an Occluder component is spawned/inserted, 16 OccluderRendererBundles are spawned, one to occlude each light source. They are despawned when the entity is despawned or the Occluder component is removed.

OccluderRendererBundle 🔒

Functions

create_occluder_polygon 🔒

despawn_collider_occluders 🔒

draw_occluders 🔒

on_remove_occluder_despawn_occluder_renderers 🔒

polygon_mesh 🔒

spawn_collider_occluders 🔒

spawn_occluder_renderers 🔒

vector_to_segment 🔒