lightborne::player
Module light

Structs

AngleMarker

PlayerLightInventory

A Component used to track Lyra’s current shooting color as well as the number of beams of that color remaining.

Functions

despawn_angle_indicator

handle_color_switch

System to handle the keyboard presses corresponding to color switches.

preview_light_path

System that uses Gizmos to preview the light path while the left mouse button is held down. This system needs some work, namely:

shoot_light

System that spawns a LightRaySource when the player releases the left mouse button. This system should instead consider sending a LightRaySpawnEvent with the needed information to keep all light-related systems in the light module.

spawn_angle_indicator