lightborne::player
Module movement

Structs

PlayerMovement

Component that stores information about the player’s movement state.

Enums

PlayerState

Constants

COYOTE_TIME_TICKS 🔒

The number of FixedUpdate steps the player can jump for after falling off an edge.

JUMP_BOOST_TICKS 🔒

The number of FixedUpdate steps the player should receive upward velocity for.

PLAYER_GRAVITY 🔒

The y velocity subtracted from the player due to gravity.

PLAYER_JUMP_VEL 🔒

The positive y velocity added to the player every jump boost tick.

PLAYER_MAX_H_VEL 🔒

Max player horizontal velocity.

PLAYER_MAX_Y_VEL 🔒

Max player vertical velocity.

PLAYER_MOVE_VEL 🔒

The x velocity added to the player when A/D is held.

SHOULD_JUMP_TICKS 🔒

The number of FixedUpdate steps the player can jump for after pressing the spacebar.

Functions

crouch_player

System that is run on Update to crouch player

move_player

System that is run on Update to move the player around.

queue_jump

System that is run the frame the space bar is pressed. Allows the player to jump for the next couple of frames.

update_player_state