lightborne::player
Module strand

Structs

Strand

Component representing one node in a chain of strands, used to simulate hair and clothes.

StrandLayerGroup

Enums

PlayerRootStrandType

Component attached to the “root” strand (the strand closest to the player, the strand with connect equal to player) for all strand chains attached to player. Used to query Strand components to updateoffset in response to player model changing, e.g. lowering hair strand when player crouches.

Functions

add_player_hair_and_cloth

add_player_strand

Creates a chain of strands to the player.

update_player_strand_offsets

System that updates Strand offsets based on PlayerRootStrandType and player state. Currently doesn’t do anything, but should be used to make Strand offsets correct when player changes direction, crouches, etc.

update_strand