lightborne::player::strand
Function add_player_strand

pub fn add_player_strand(
    strand_dist: f32,
    strand_gravity: Range<f32>,
    strand_friction: f32,
    layer_groups: &[StrandLayerGroup<'_>],
    layer_group_order: &[(usize, bool)],
    player_offset: Vec3,
    player_root_strand_type: PlayerRootStrandType,
    commands: &mut Commands<'_, '_>,
    player_entity: Entity,
    asset_server: &Res<'_, AssetServer>,
    sprite_translate: Vec2,
)

Creates a chain of strands to the player.

Each created Strand component has a dist of strand_dist, and a gravity of at strand_gravity.start near the player that slowly turns into strand_gravity.end. The function layers the sprites in each list of layer_groups based on their order, and creates an entity for each index in layer_group_order.