lightborne::player::strand
Function update_player_strand_offsets

pub fn update_player_strand_offsets(
    strands: Query<'_, '_, (&mut Strand, &PlayerRootStrandType)>,
    player: Query<'_, '_, (), With<PlayerMarker>>,
)

System that updates Strand offsets based on PlayerRootStrandType and player state. Currently doesn’t do anything, but should be used to make Strand offsets correct when player changes direction, crouches, etc.