lightborne::light
Module sensor

Structs

HitByLightEvent

Event used to notify other entities to trigger based on collision with light. An included Entity is used to indicate the corresponding LightSensor.

LightSensor

Component added to entities receptive to light. The activation_timer should be initialized in the From<&EntityInstance> implemenation for the LightSensorBundle, if not default.

LightSensorBundle

Bundle that includes all the Components needed for a LightSensor to function properly.

Functions

reset_light_sensors

System that resets the LightSensors when a [LevelSwitchEvent] is received.

update_light_sensors

System that runs on Update, querying each LightSensor and updating them based on each HitByLightEvent generated in the System: simulate_light_sources. This design is still imperfect, as while it differs semantically from the previous implementation, each Event is generated every frame. Preferably, refactor to include a “yap”-free implementation across multiple systems to better utilize Event.